export de la logique des route user et mise en place doc SWAGGER
