+++ middleware/private.js
@@ -1,14 +1,20 @@
-/**
- * * @param {Object} req - L'objet de requête Express
- * * @param {Object} res - L'objet de réponse Express
- * * @param {Function} next - La fonction middleware suivante dans la pile
- * Middleware pour restreindre l'acces au route aux utilisateurs authentifiés
- */
+const jwt = require('jsonwebtoken');
+const SECRET_KEY = process.env.SECRET_KEY;
+const C = require('../test/test');
 
-module.exports = function (req, res, next) {
-    if (req.session && req.session.user) {
-        next(); 
-    } else {
-        res.status(401).json({ message: 'Unauthorized' });
-    }   
-};
+exports.checkJWT = (req, res, next) => {
+    const token = req.session.token;
+    if (!token) {
+        C.log('red', 'Token non fourni en session');
+        return res.status(401).redirect('/');
+    }
+    jwt.verify(token, SECRET_KEY, (err, decoded) => {
+        if (err) {
+            C.log('red', 'Token JWT invalide');
+            return res.status(401).json('token_not_valid');
+        } else {
+            req.decoded = decoded;
+            next();
+        }
+    });
+}
