+++ routes/login.js
@@ -2,28 +2,26 @@ const express = require('express');
 const router = express.Router();
 
 const C = require('../test/test');
-const verifUsers = require('../test/verifUsers');
+const verifUsers = require('../middleware/verifUsers');
 
-router.post('/', async (req,res,next) => {
-    //récup des variables
-    // const visit = req.body.visit;  
-    // const email = req.body.email;
-    // const password = req.body.password;
-    // const name = 'bruno'
-
+router.post('/', async (req,res,) => {
 
     let validUser = await verifUsers.verifUsers(req.body.email, req.body.password);
 
 
     if(validUser) {
+        const { user, token } = validUser;
          // mise en session
         const session = req.session;
         session.verif = true;
-        session.email = validUser.email;
-        session.name = validUser.name;
+        session.email = user.email;
+        session.name = user.name;
         session.visite = req.body.visit;
-
+        session.token = token;
         delete session.errMsg;
+        
+        //res.header('Authorization', 'Bearer ' + token);
+        C.log('green' `Juste apres le token username  => ${user.name}`)
 
         res.status(301).redirect('/dashboard')
